Second module completed
The basic concepts of python were covereed in this two modules.
The second module focuses on data types and data structures, likes strings, lists, dictionaries and tuples.
In many exercises, it was required to read the content from a text file, thats why many excercises are organized iun folders.
All the assignments where scripted, tested and submitted to both the course platform and this repository.