Third module completed
This module focuses on extracting data from web sources. The low level connection was introduced with Socket library,
then high level conection was done with urllib. The direct HTML scrapping was studied with BeautifulSoup library,
the XML format was handled as a tree with ElementTree and finally the json format was handled by dictionaries.
Furthermore, basic API concepts were introduced.
All the assignments where scripted, tested and submitted to both the course platform and this repository.